以下の関数を修正
・parse_number_value()
・json_value_init_number()
・json_value_get_number()

+++ pspplugin/parson/parson.h
@@ -258,7 +258,7 @@ JSON_Value * json_value_init_string (const char *string); /* copies passed strin
 JSON_Value * json_value_init_string_with_len(const char *string, size_t length); /* copies passed string, length shouldn't include last null character */
 JSON_Value * json_value_init_number(
     SceUID *    valueIdPtr
-    , double    number
+    , long      number
 );
 JSON_Value * json_value_init_boolean(
     SceUID *    valueIdPtr
@@ -278,7 +278,7 @@ JSON_Object *   json_value_get_object (const JSON_Value *value);
 JSON_Array  *   json_value_get_array  (const JSON_Value *value);
 const char  *   json_value_get_string (const JSON_Value *value);
 size_t          json_value_get_string_len(const JSON_Value *value); /* doesn't account for last null character */
-double          json_value_get_number (const JSON_Value *value);
+long            json_value_get_number (const JSON_Value *value);
 int             json_value_get_boolean(const JSON_Value *value);
 JSON_Value  *   json_value_get_parent (const JSON_Value *value);
